+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Change log
 ==========
 
-not released
+v0.1.0a11
 ------------
 
 - The ``filtering.rpsl`` section of general and clients configuration files has been renamed into ``filtering.irrdb``.